chore: remove all old copilot dashboard components and files
