+++ frontend/src/components/Copilot/Breakdowns/SuggestionsAcceptancesSizeComparisonGraph.js
